requirements/base.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.7
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
drf-yasg==1.10.0
flex==6.13.2
future==0.16.0
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
MarkupSafe==1.0
mock==2.0.0
openapi-codec==1.3.2
pbr==4.1.0
psycopg2==2.7.5
pytz==2018.5
PyYAML==3.13
requests==2.19.1
rfc3987==1.3.8
ruamel.yaml==0.15.59
simplejson==3.16.0
six==1.11.0
strict-rfc3339==0.7
swagger-spec-validator==2.3.1
uritemplate==3.0.0
urllib3==1.23
validate-email==1.3
